Refactor init command and create InitManager class
- Move init logic to new InitManager class
- Implement error handling and retry mechanism for AI response
- Add directory existence check for generated config
- Improve code organization and readability
